requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
Brlapi==0.6.6
certifi==2019.6.16
chardet==3.0.4
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.2.4
django-csvimport==2.12
django-filter==2.2.0
django-filters==0.2.1
django-pandas==0.6.1
django-postgres-copy==2.4.1
djangorestframework==3.10.2
httplib2==0.9.2
idna==2.6
isort==4.3.21
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
louis==3.5.0
macaroonbakery==1.1.3
mccabe==0.6.1
numpy==1.17.0
oauth==1.0.1
olefile==0.45.1
pandas==0.25.0
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
protobuf==3.0.0
psycopg2==2.8.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.3
virtualenv-clone==0.5.3
wadllib==1.3.2
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
